MQTTBridge/requirements.txt
click==6.7
paho-mqtt==1.3.0
pip==1.5.6
python-dateutil==2.6.1
PyYAML==3.12
rethinkdb==2.3.0.post6
setuptools==2.1
six==1.11.0
